requirements.txt
django
django-import-export
django-suit
django_select2